refactor: refactor attachment list
